Functions moved to helper folder
