+++ src/common/helper/profileUpdate/configFiles/vcArray.json
