+++ src/common/helper/profileUpdate/vcPaths/casteCertificate.json
@@ -1,0 +1,3 @@
+{
+  "caste": "credentialSubject.castCategoryNumber"
+}
