+++ src/common/helper/profileUpdate/vcPaths/enrollmentCertificate.json
