--- src/modules/users/crons/vcPaths/casteCertificate.json
@@ -1,3 +1,0 @@
-{
-  "caste": "credentialSubject.castName"
-}
